[tf.data] move csv checkpoint tests to kernel tests
